Breaks resolvers apart into multiple files roughly organized around graphql types or logical groups of those types.
